templates/attested-api/python/requirements.txt
eth-account==0.13.7
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn[standard]==0.30.6
